// src/module/users/dto/create-user.dto.ts
export interface ProvisionedAccountsDTO {
  palmpay: {
    accountId?: string;
    accountNumber?: string;
    accountName?: string;
    bankName: string;
    currency: string;
  };
  pouchii: {
    walletId: number;
    accountNumber: string;
    accountName: string;
    bankName: string;
    currency: string;
  };
}